policies
Policies: Product, TrainingType and Room

+++ app/Http/Controllers/TrainingTypeController.php
@@ -5,16 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\TrainingType;
 use App\Http\Requests\StoreTrainingTypeRequest;
 use App\Http\Requests\UpdateTrainingTypeRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class TrainingTypeController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $training_types = TrainingType::orderBy('id', 'desc')->paginate(15);
-        return view('pages.training_types.index', ['training_types' => $training_types]);
+        $this->authorize('viewAny', TrainingType::class);
+        $trainingTypes = TrainingType::all();
+        return view('pages.training_types.index', ['training_types' => $trainingTypes]);
     }
 
     /**
@@ -22,6 +25,7 @@ class TrainingTypeController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', TrainingType::class);
         return view('pages.training_types.create');
     }
 
@@ -31,7 +35,14 @@ class TrainingTypeController extends Controller
     public function store(StoreTrainingTypeRequest $request)
     {
         $validatedData = $request->validated();
+
+        if ($request->hasFile('image')) {
+            $path = $request->file('image')->store('training_types', 'public');
+            $validatedData['image'] = $path;
+        }
+
         TrainingType::create($validatedData);
+
         return redirect()->route('training_types.index')->with('success', 'Training type created successfully.');
     }
 
@@ -40,6 +51,7 @@ class TrainingTypeController extends Controller
      */
     public function show(TrainingType $trainingType)
     {
+        $this->authorize('view', $trainingType);
         return view('pages.training_types.show', ['training_type' => $trainingType]);
     }
 
@@ -48,6 +60,7 @@ class TrainingTypeController extends Controller
      */
     public function edit(TrainingType $trainingType)
     {
+        $this->authorize('update', $trainingType);
         return view('pages.training_types.edit', ['training_type' => $trainingType]);
     }
 
@@ -57,7 +70,14 @@ class TrainingTypeController extends Controller
     public function update(UpdateTrainingTypeRequest $request, TrainingType $trainingType)
     {
         $validatedData = $request->validated();
+
+        if ($request->hasFile('image')) {
+            $path = $request->file('image')->store('training_types', 'public');
+            $validatedData['image'] = $path;
+        }
+
         $trainingType->update($validatedData);
+
         return redirect()->route('training_types.index')->with('success', 'Training type updated successfully.');
     }
 
@@ -66,6 +86,7 @@ class TrainingTypeController extends Controller
      */
     public function destroy(TrainingType $trainingType)
     {
+        $this->authorize('delete', $trainingType);
         $trainingType->delete();
         return redirect()->route('training_types.index')->with('success', 'Training type deleted successfully.');
     }
